+++ docs/versionhistory.rst
@@ -8,6 +8,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - Fixed broken ``MultiListener.extra()`` (PR by daa)
 - Fixed ``TLSStream`` returning an empty bytes object instead of raising ``EndOfStream`` when
   trying to receive from the stream after a closing handshake
+- Fixed ``AttributeError`` when cancelling a task group's scope inside an async test fixture on
+  asyncio
 
 **2.0.0**